Create README.md
We are thrilled to announce and celebrate the incredible collaboration between @energiasocial, a pioneering renewable social business in LATAM, and @txpipe_tools , an OpenSource tools and infrastructure provider for Cardano blockchain developers. This partnership marks a significant milestone in the integration of on-chain data with off-chain infrastructure, bringing about a positive change in the world of renewable energy and blockchain technology.

@energiasocial has long been at the forefront of the renewable energy revolution, committed to harnessing the power of clean and sustainable resources to drive social impact. With their unwavering dedication to creating a better world, they have captured the hearts of many and inspired change on a global scale.

@TxPipe, on the other hand, has emerged as a trailblazer in the world of blockchain development, focusing on providing tools and infrastructure that empower developers on the Cardano platform. Their commitment to open-source principles and collaboration has earned them a reputation as a trusted partner for developers worldwide. 

By joining forces, @energiasocial and @TxPipe are set to revolutionize the integration of on-chain data with off-chain infrastructure. This collaboration will enable renewable energy projects to leverage the transparency and security of blockchain technology, facilitating efficient and sustainable growth.